Split the Drawable into 2 interfaces
