+++ core/src/com/po/fuck/FUCK.java
@@ -1,7 +1,5 @@
 package com.po.fuck;
 
-import java.util.ArrayList;
-
 import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Gdx;
 import com.po.fuck.view.Renderer;
@@ -23,11 +21,6 @@ public class FUCK extends ApplicationAdapter {
     public void render() {
         float delta = Gdx.graphics.getDeltaTime();
         core.update(delta);
-        ArrayList<Object> drawables = new ArrayList<>();
-        for(Object obj : core.getDrawableCollection()){
-            drawables.add(obj);
-        }
-        drawables.add(Core.coinsCounter.get());
-        renderer.render(drawables);
+        renderer.render(core.getDrawableCollection());
     }
 }
