+++ core/src/com/po/fuck/model/GameObject.java
@@ -3,6 +3,7 @@ package com.po.fuck.model;
 import com.badlogic.gdx.math.Polygon;
 import com.badlogic.gdx.math.Vector2;
 import com.po.fuck.model.collision.Collidable;
+import com.po.fuck.model.drawables.Drawable;
 import com.po.fuck.model.lifetime.Destructable;
 import com.po.fuck.model.position.GeometryData;
