+++ core/src/com/po/fuck/model/drawables/BasicDrawable.java
@@ -1,0 +1,7 @@
+package com.po.fuck.model.drawables;
+
+public interface BasicDrawable {
+    default int get_z() {
+        return 0;
+    }
+}
