+++ core/src/com/po/fuck/model/drawables/Drawable.java
@@ -1,13 +1,12 @@
-package com.po.fuck.model;
+package com.po.fuck.model.drawables;
 
 import com.badlogic.gdx.math.Vector2;
 import com.po.fuck.model.position.GeometryData;
 
-public interface Drawable {
-    default int get_z() {
-        return 0;
-    }
+public interface Drawable extends BasicDrawable {
+
     public Vector2 getPosition();
+
     public GeometryData getGeometryData();
 }
