+++ core/src/com/po/fuck/model/weapons/Weapon.java
@@ -4,8 +4,8 @@ import static com.po.fuck.model.Constants.WEAPON_LAYER;
 
 import com.badlogic.gdx.math.Vector2;
 import com.po.fuck.model.Entity;
+import com.po.fuck.model.drawables.Drawable;
 import com.po.fuck.model.position.GeometryData;
-import com.po.fuck.model.Drawable;
 
 public abstract class Weapon implements Drawable {
     protected final Entity owner;
